[PRJ][APPS][USBH] Fix stack corruption due to inconsistent type of local variable in MSC_File_Operations()

+++ Projects/STM32H743I-EVAL/Applications/USB_Host/DualCore_Standalone/Src/file_operations.c
@@ -40,7 +40,7 @@ uint8_t wtext[] = "USB Host Library : Mass Storage application";
   */
 void MSC_File_Operations(void)
 {
-  uint16_t bytesread;
+  uint32_t bytesread;
 
   /* Register the file system object to the FatFs module */
   if(f_mount(&USBH_fatfs, "", 0) != FR_OK)
